requirements.txt
black==25.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
et_xmlfile==2.0.0
flake8==7.3.0
google-api-core==2.25.2
google-api-python-client==2.184.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
httplib2==0.31.0
idna==3.10
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mccabe==0.7.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
mypy==1.18.2
mypy_extensions==1.1.0
oauthlib==3.3.1
openpyxl==3.1.5
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
proto-plus==1.26.1
protobuf==6.32.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pyparsing==3.2.5
pytokens==0.1.10
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
textual==6.1.0
typing_extensions==4.15.0
uc-micro-py==1.0.3
uritemplate==4.2.0
urllib3==2.5.0
